feature(WIP): restricing access

+++ src/pages/api/folder-tags/[id].ts
@@ -7,7 +7,7 @@ export default async function folderHandler(req: NextApiRequest, res: NextApiRes
 
     const user = await verifySessionInApi(req, res);
 
-	if (!user) {return;}
+    if (!user) { return; }
 
     await db.query(`
                 CREATE TABLE IF NOT EXISTS Folder (
@@ -21,14 +21,23 @@ export default async function folderHandler(req: NextApiRequest, res: NextApiRes
                 );
             `);
 
+    const folderOwner = await db.query('SELECT osoba.id FROM osoba JOIN folder ON osoba.id = folder.osoba WHERE folder.id = $1', [query.id]);
+
+    const isAllowedToEditTags = (await db.query(
+        `SELECT * FROM dostep WHERE folder = $1 AND osoba = $2 AND edycja = TRUE`, [query.id, user.id]
+    )).rows.length > 0;
+
     if (method === 'POST') {
         // Create a folder
         const { klucz, wartosc } = body;
         try {
-
+            if (!isAllowedToEditTags && folderOwner.rows[0].osoba !== user.id && (!user.administrator || !user.edytowanieFolderow)) {
+                res.status(401).json({ message: "Unauthorized" });
+                return;
+            }
             const result = await db.query(
                 `INSERT INTO TagFolderu (idfolderu, klucz, wartosc) VALUES ($1, $2, $3) RETURNING *;`,
-                [ query.id, klucz, wartosc,]
+                [query.id, klucz, wartosc,]
             );
 
             res.status(201).json(result.rows[0]);
@@ -46,7 +55,16 @@ export default async function folderHandler(req: NextApiRequest, res: NextApiRes
             //     PRIMARY KEY (idFolderu, klucz),
             //     FOREIGN KEY (idFolderu) REFERENCES Folder(id)
             // );
-            
+            const isAllowed = (await db.query(
+                `SELECT * FROM dostep WHERE folder = $1 AND osoba = $2`, [query.id, user.id]
+            )).rows.length > 0;
+
+            if (!isAllowed && folderOwner.rows[0].osoba !== user.id && !user.administrator) {
+                res.status(401).json({ message: "Unauthorized" });
+                return;
+            }
+
+
             const result = await db.query(`SELECT klucz, wartosc FROM TagFolderu WHERE idFolderu = $1`, [query.id]);
             res.status(200).json(result.rows);
         } catch (error) {
@@ -55,6 +73,10 @@ export default async function folderHandler(req: NextApiRequest, res: NextApiRes
         }
     } else if (method === 'DELETE') {
         try {
+            if (!isAllowedToEditTags && folderOwner.rows[0].osoba !== user.id && (!user.administrator || !user.edytowanieFolderow)) {
+                res.status(401).json({ message: "Unauthorized" });
+                return;
+            }
             await db.query(`DELETE FROM TagFolderu WHERE idfolderu = $1 AND klucz = $2`, [query.id, body.klucz]);
             res.status(200).json({ message: "Folder tag deleted successfully" });
         } catch (error) {
@@ -64,6 +86,10 @@ export default async function folderHandler(req: NextApiRequest, res: NextApiRes
 
     } else if (method === 'PUT') {
         try {
+            if (!isAllowedToEditTags && folderOwner.rows[0].osoba !== user.id && (!user.administrator || !user.edytowanieFolderow)) {
+                res.status(401).json({ message: "Unauthorized" });
+                return;
+            }
             const { klucz, wartosc } = req.body;
             await db.query(`UPDATE TagFolderu SET wartosc = $2 WHERE idFolderu = $3 AND klucz = $1`, [
                 klucz,
